Поменяла структуру проекта согласно MVVM
