+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AmmyMovie"
         tools:targetApi="m">
-        <activity android:name=".ui.main.view.MainActivity">
+        <activity android:name="com.example.ammymovie.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
