+++ app/src/main/java/com/example/ammymovie/MainActivity.kt
@@ -1,24 +1,16 @@
-package com.example.ammymovie.ui.main.view
+package com.example.ammymovie
 
 import android.os.Build
 import android.os.Bundle
-import android.os.Handler
-import android.util.Log
 import android.view.Menu
 import android.widget.Toast
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.SearchView
 import androidx.fragment.app.Fragment
-import com.example.ammymovie.R
 import com.example.ammymovie.databinding.MainActivityBinding
+import com.example.ammymovie.ui.main.MainFragment
 import kotlinx.android.synthetic.main.main_activity.*
-import java.io.BufferedReader
-import java.io.InputStreamReader
-import java.net.MalformedURLException
-import java.net.URL
-import java.util.stream.Collectors
-import javax.net.ssl.HttpsURLConnection
 
 class MainActivity : AppCompatActivity() {
