+++ app/src/main/java/com/example/ammymovie/domain/repository/RepositoryImpl.kt
@@ -1,4 +1,7 @@
-package com.example.ammymovie.ui.main.model
+package com.example.ammymovie.domain.repository
+
+import com.example.ammymovie.domain.model.getNowPlayingMovies
+import com.example.ammymovie.domain.model.getUpcomingMovies
 
 class RepositoryImpl : Repository {
     override fun getNowPlayingFromLocalStorage() = getNowPlayingMovies()
