+++ app/src/main/java/com/example/ammymovie/ui/detail/DitailsViewModel.kt
@@ -1,0 +1,10 @@
+package com.example.ammymovie.ui.detail
+
+/**
+ * homework com.example.ammymovie.ui.ditail.viewmodel
+ *
+ * @author Amina
+ * 23.07.2021
+ */
+class DitailsViewModel {
+}
