+++ app/src/main/java/com/example/ammymovie/ui/main/NowPlayingFragmentAdapter.kt
@@ -1,12 +1,13 @@
-package com.example.ammymovie.ui.main.view
+package com.example.ammymovie.ui.main
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.ammymovie.databinding.ItemNowPlayingBinding
-import com.example.ammymovie.ui.main.model.Movie
-
+import com.example.ammymovie.domain.model.Movie
+import com.example.ammymovie.utils.format
+import com.example.ammymovie.view.createImageFromResource
 
 class NowPlayingFragmentAdapter : RecyclerView.Adapter<NowPlayingFragmentAdapter.ViewHolder>() {
